pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lang-ipea-editorial"
version = "0.2.0"
description = "Chat com DOCX/PDF e revisao editorial com Streamlit + LangGraph"
readme = "Readme.md"
requires-python = ">=3.10"
dependencies = [
  "lxml>=5.2.0",
  "langchain>=1.0.0",
  "langgraph>=1.0.0",
  "langchain-openai>=1.0.0",
  "python-dotenv>=1.0.0",
  "streamlit[pdf]>=1.55.0",
  "pypdf>=5.0.0",
  "pymupdf>=1.24.0",
  "pydantic>=2.7.0"
]

[tool.setuptools.packages.find]
where = ["src"]